// bot/src/interfaces/Hivemind.interface.ts
interface PermissionsBitField {
    bitfield?: number
    allPermissions?: number
    defaultPermissions?: number
    permissionFlags?: Record<string, boolean>
    stageModerator?: number
}
// ...
interface AllowedMentions {
    parse?: string[]
    roles?: number[]
    users?: string[]
    replied_user?: boolean
}

// Embed
interface Attachment {
    id: string
    filename?: string
    description?: string
    content_type?: string
    size?: number
    url?: string
    proxy_url?: string
    height?: number
    width?: number
    ephemeral?: boolean
    duration_secs?: number
    waveform?: string
    flags?: number
}

// Embed
interface Embed {
    title?: string
    type?: string
    description?: string
    url?: string
    timestamp?: string
    color?: number
    footer?: Record<string, any>
    image?: Record<string, any>
    thumbnail?: Record<string, any>
    video?: Record<string, any>
    provider?: Record<string, any>
    author?: Record<string, any>
    fields?: Array<Record<string, any>>
}

// InteractionCallbackData
interface InteractionCallbackData {
    tts?: boolean | null
    content?: string | null
    embeds?: Embed[] | null
    allowed_mentions?: string | null
    flags?: number | null
    components?: Array<Record<string, any>> | null
    attachments?: Attachment[] | null
}

// ChatInputCommandInteraction
export interface ChatInputCommandInteraction_broker {
    id: string
    applicationId?: string | null
    type?: number | null
    channel?: Record<string, any> | null // ?
    channelId?: string | null
    token?: string | null
    guildId?: string | null
    user?: Record<string, any> | null
    createdAt?: Date | null
    deferred?: boolean | null
    replied?: boolean | null
    webhook?: Record<string, any> | null
    member?: Record<string, any> | null
    ephemeral?: boolean | null
    guild?: Record<string, any> | null // ?
    createdTimestamp?: number | null
    appPermissions?: PermissionsBitField | null
    locale?: string | null
    guildLocale?: string | null
    client?: Record<string, any> | null
    command?: Record<string, any> | null
    commandId?: string | null
    commandName?: string | null
    commandType?: any | null
    commandGuildId?: string | null
    memberPermissions?: PermissionsBitField | null
    options?: Record<string, any> | null
    version?: number | null
}

// FollowUpMessageData
export interface FollowUpMessageData {
    content?: string
    username?: string
    avatar_url?: string
    tts?: boolean
    embeds?: Embed[]
    allowed_mentions?: AllowedMentions
    components?: any[]
    files?: any[] //
    payload_json?: string
    attachments?: Attachment[]
    flags?: number
    thread_name?: string
}

// InteractionResponseEditData
export interface InteractionResponseEditData {
    thread_id?: number | null
    content?: string | null
    embeds?: Embed[] | null
    allowed_mentions?: AllowedMentions | null
    components?: Array<Array<Record<string, any>>> | null
    files?: Array<Record<string, any>> | null
    payload_json?: string | null
    attachments?: Attachment[] | null
}

// InteractionResponse
export interface InteractionResponse {
    type: number
    data?: InteractionCallbackData | null
}